Present authorized scopes to user, post-auth
Previously, the scopes that were presented to the user (alongside their
`me`) were the scopes that Micropublish requested, rather than the
scopes returned by the authorization server.

We should make sure we present those that were actually returned by the
token endpoint.

To match the intent, we should update our methods to note that they also
return the `scope`, and make sure we update our session object with it.